refactor(library/URLComponent): moves "Origin.ts" to "Origin/definition.ts"
