+++ src/library/URLComponent/Origin/definition.ts
